update: remove utils in prompt editor
